feat: use piscina
