+++ packages/native-federation-typescript/package.json
@@ -56,7 +56,8 @@
     "ansi-colors": "^4.1.3",
     "axios": "^1.6.7",
     "rambda": "^9.1.0",
-    "unplugin": "^1.9.0"
+    "unplugin": "^1.9.0",
+    "piscina": "^4.4.0"
   },
   "peerDependencies": {
     "typescript": "^4.9.0 || ^5.0.0",
