+++ packages/native-federation-typescript/src/index.ts
@@ -1,5 +1,4 @@
 import ansiColors from 'ansi-colors';
-import { rm } from 'fs/promises';
 import { resolve } from 'path';
 import { mergeDeepRight } from 'rambda';
 import { UnpluginOptions, createUnplugin } from 'unplugin';
@@ -8,31 +7,22 @@ import { retrieveHostConfig } from './configurations/hostPlugin';
 import { retrieveRemoteConfig } from './configurations/remotePlugin';
 import { HostOptions } from './interfaces/HostOptions';
 import { RemoteOptions } from './interfaces/RemoteOptions';
-import { createTypesArchive, downloadTypesArchive } from './lib/archiveHandler';
-import {
-  compileTs,
-  retrieveMfTypesPath,
-  retrieveOriginalOutDir,
-} from './lib/typeScriptCompiler';
+import { retrieveOriginalOutDir } from './lib/typeScriptCompiler';
+// @ts-expect-error function exported through module.exports
+import remoteWriteBundle from './lib/writeBundle/remote';
+// @ts-expect-error function exported through module.exports
+import hostWriteBundle from './lib/writeBundle/host';
 
 export const NativeFederationTypeScriptRemote = createUnplugin(
   (options: RemoteOptions) => {
-    const { remoteOptions, tsConfig, mapComponentsToExpose } =
-      retrieveRemoteConfig(options);
+    const remoteConfig = retrieveRemoteConfig(options);
+    const { remoteOptions, tsConfig } = remoteConfig
     return {
       name: 'native-federation-typescript/remote',
       async writeBundle() {
         try {
-          compileTs(mapComponentsToExpose, tsConfig, remoteOptions);
+          await remoteWriteBundle(remoteConfig)
 
-          await createTypesArchive(tsConfig, remoteOptions);
-
-          if (remoteOptions.deleteTypesFolder) {
-            await rm(retrieveMfTypesPath(tsConfig, remoteOptions), {
-              recursive: true,
-              force: true,
-            });
-          }
           console.log(ansiColors.green('Federated types created correctly'));
         } catch (error) {
           console.error(
@@ -44,9 +34,9 @@ export const NativeFederationTypeScriptRemote = createUnplugin(
         return process.env.NODE_ENV === 'production'
           ? undefined
           : {
-              buildStart: (this as UnpluginOptions).writeBundle,
-              watchChange: (this as UnpluginOptions).writeBundle,
-            };
+            buildStart: (this as UnpluginOptions).writeBundle,
+            watchChange: (this as UnpluginOptions).writeBundle,
+          };
       },
       webpack(compiler) {
         compiler.options.devServer = mergeDeepRight(
@@ -78,35 +68,21 @@ export const NativeFederationTypeScriptRemote = createUnplugin(
 
 export const NativeFederationTypeScriptHost = createUnplugin(
   (options: HostOptions) => {
-    const { hostOptions, mapRemotesToDownload } = retrieveHostConfig(options);
+    const hostConfig = retrieveHostConfig(options);
     return {
       name: 'native-federation-typescript/host',
       async writeBundle() {
-        if (hostOptions.deleteTypesFolder) {
-          await rm(hostOptions.typesFolder, {
-            recursive: true,
-            force: true,
-          }).catch((error) =>
-            console.error(
-              ansiColors.red(`Unable to remove types folder, ${error}`),
-            ),
-          );
-        }
+        await hostWriteBundle(hostConfig);
 
-        const typesDownloader = downloadTypesArchive(hostOptions);
-        const downloadPromises =
-          Object.entries(mapRemotesToDownload).map(typesDownloader);
-
-        await Promise.allSettled(downloadPromises);
         console.log(ansiColors.green('Federated types extraction completed'));
       },
       get vite() {
         return process.env.NODE_ENV === 'production'
           ? undefined
           : {
-              buildStart: (this as UnpluginOptions).writeBundle,
-              watchChange: (this as UnpluginOptions).writeBundle,
-            };
+            buildStart: (this as UnpluginOptions).writeBundle,
+            watchChange: (this as UnpluginOptions).writeBundle,
+          };
       },
     };
   },
